reorganized_project/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
sqlalchemy>=2.0.15
pymongo>=4.3.3
pyjwt>=2.7.0
cryptography>=40.0.2
passlib>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
aiofiles>=23.1.0

# AI and machine learning
torch>=2.0.0
transformers>=4.29.2
numpy>=1.24.3
pandas>=2.0.2
scikit-learn>=1.2.2
